change movement to be up/right/up/down instead of forward/backward
Thanks to Jingying for play-testing and pointing out how confusing
it was before. Note that this does make defining spells a bit harder
since the spells will be fixed to a particular direction. We could
probably find a way to rotate all of the actions in a spell based
on its the relative orientation of the player between definition and
casting.