added a level "Glider track"
This level is interesting because it breaks the "relative reachability"
side effect metric proposed by Vika. Once you stop the glider it's
impossible to get to any state that was reachable from the baseline
where the glider was *not* stopped. However, if the baseline is taken
to be the initial state rather than the inaction state, then it should
still work out fine.