get rid of old game.py; rename src folder to 'safety_net'
The only thing that was in the old `game.py` file that is no longer
present in the other files is the ability to run the game using
absolute coordinates. I didn't really like absolute coordinates (even
though others found the relative coordinates to be confusing), and
they didn't mesh well with zero-cost rotation actions.