allow both absolute and relative directions at the same time
I'm now making a distinction between FORWARD and UP. Agents can now
toggle cells in a particular direction. This should make it easier to
represent the current state, because the agent's orientation no longer
matters. It does make the action space slightly larger (4 directions
with and without toggle, plus null), but since I've condensed all
agent actions to "toggle" it's not so bad.