minor build improvements
I'm renaming the speedups source directory from `speedups` to
`speedups_src`. This makes it so that there's not a name conflict
between the source files and the actual compiled extension module, so
when the extension module isn't built properly the associated error
is much more informative.

I'm also creating a setup.cfg file that makes it easier to run
`setup.py build` and flake8.