get rid of formatted string literals; set required python version = 3.5
(formatted string literals were introduced in python 3.6)

I've also changed the error behavior on old python versions to be an
ImportError for Python < 3 and a warning for Python < 3.5. This is
more flexible than a print statement, because it allows anyone that's
importing SafeLife as a module as part of another package to handle or
ignore the error / warning in whatever way they see fit.