better side effects scores and logging
When I actually went to measure the side effects for some of my agents,
I realized that the side effect scores weren't that useful. Problem is
that I really want to see a distribution of side effects per agent,
not a single number which measures the side effects for a distribution
of agents. The latter could easily mask some very unsafe behavior.