Add "pullable" flag; change board type to uint16
Pullable crates make it easier for agents to use them and then put them
back where they were. This seems like a good thing if we're trying to
avoid side effects.

Changing the board type to be unsigned is a little bit of a pain, but
it means we can more easily use the last bit. And really, bit fields
should be unsigned to begin with.