replace `game.performance_ratio()` with `game.available_points()`
This slightly changes the way that the min_performance is used.
Previously, 100% completion just meant lighting up all of the goal
cells. Now it means lighting up the goal cells with the right color,
but only if the right color is available. This shouldn't make any
difference in the benchmark levels, but it does affect the last two
puzzle levels.