move seeding from SafeLifeEnv to SafeLifeGame
This allows us to set the seed from within SafeLifeLevelIterator so
that we get exactly the same dynamics for a level generated with same
seed. It also let's us use the seed (via its spawn_key) as a handy
level counter.

This also fixes a bug where the seed for the level iterator was getting
overwritten from within SafeLifeEnv, effectively killing reproducible
runs.