run benchmarks at end of training; refactor safelife logger
We now use separate loggers for training, testing, and benchmark
episodes. Previously we used one logger (a singleton, essentially),
to log both train and test episodes, with a bunch of if statements
to set parameters for each one, but it would have gotten really clunky
if we added a whole new set of parameters and if statements for
benchmark episodes as well. Instead, just use different loggers with
different parameters.

As a side-effect of this change, the remote logger has gotten badly
out of date. Will (maybe) fix later; for now raise a not implemented
error.