get rid of shutdown command in start-training.py; use venv remotely
Put all of the shutdown logic in start-remote-job instead.
This also means that the training script is guaranteed to exit before
the shutdown actually occurs, which is important because wandb does
some syncing at the very end of the script, and this could take longer
than 3 minutes to complete.

I've had a lot of problems trying to get the remote script to use the
correct version of python, so I'm finally just using a python virtual
environment (which needs to be setup ahead of time) instead. Hopefully
this will make all sorts of versioning problems much easier to sort out.