more improvements to logging and start-remote-job
The git remote of the cloned repo on the remote job should now match
the remote of the current repo (i.e., point to github). This means
wandb can point to a real github commit id.

I've renamed a bunch of the summary statistics to make them a little
easier to understand and easier to read in the wandb benchmark table.

I've also added env-type to the summary statistics so that it too can
be displayed in the benchmark table.