Set benchmark 'steps' stat to include avg from unsuccessful episodes
I was waffling on this. I now think in the graph it's easier to see
what's going on if the length is only logged for those cases where
the agent actually completes the episode, but for the benchmark
statistic it's just confusing to have it be anything other than the
average length. You can always convert between one and the other using

	avg_steps = success * successful_steps + (1-success) * 1000