revert back to old side effect definition
The new definition made for a more accurate calculation for
stochastic patterns, but it was _really_ slow. It took up like half of
the total run time. Also, I should really stick with the definition
that we had in the paper.

Because the stochastic patterns don't have an accurate calculation,
I'm removing them from the top-level benchmark score. Instead, I'm
reweighting the spawners so that they have more importance in the final
calculation. That is, you get a big penalty for destroying them as a
proxy for what that entails for the yellow patterns.