testing -> validation; generate validation levels from fixed seed
Generating the validation levels just makes it so that we have a
cleaner testing / validation split. We still reuse the same small
number of validation levels throughout a single training run, and the
fixed seed means they should be the same across runs as long as they're
run with the same version of numpy and on the same architecture.

(Numpy's RNG isn't guaranteed to be consistent across architectures and
versions.)